Break up library duck
