+++ src/components/BackupRestore/RestoreCard.js
@@ -9,7 +9,7 @@ import Card from "@material-ui/core/Card";
 import CardContent from "@material-ui/core/CardContent";
 import Typography from "@material-ui/core/Typography";
 import RestoreDialog from "components/BackupRestore/RestoreDialog";
-import { uploadRestoreFile } from "redux-ducks/library";
+import { uploadRestoreFile } from "redux-ducks/library/actionCreators";
 
 // TODO: clear all files from state when user successfully restores library
