+++ src/redux-ducks/library/libraryUtils.js
